Add budget tracker as separate page from calculator
